# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sons of Aral Website",
        "description": "Documentación de la API para Sons of Aral Website",
        "version": "1.0"
    },
    "paths": {
...
        "/api/canciones": {
            "get": {
                "tags": [
                    "Canciones"
                ],
                "summary": "Obtener todas las canciones",
                "operationId": "6da63c0dd890173a5268f67bfd7c1740",
                "responses": {
                    "200": {
                        "description": "Lista de canciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "lanzamientoId": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "lanzamiento": {
                                                        "type": "string",
                                                        "example": "Nombre del lanzamiento",
                                                        "nullable": true
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Mi canción"
                                                    },
                                                    "duracion": {
                                                        "description": "Duración en segundos",
                                                        "type": "integer",
                                                        "example": 225
                                                    },
                                                    "duracionFormateada": {
                                                        "type": "string",
                                                        "example": "03:45"
                                                    },
                                                    "trackNumber": {
                                                        "type": "integer",
                                                        "example": 1,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Canciones"
                ],
                "summary": "Crear una canción",
                "operationId": "5d96d76e71f35ab29f20a760271c4757",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "duracion",
                                    "lanzamiento_id"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Mi canción"
                                    },
                                    "duracion": {
                                        "type": "integer",
                                        "example": 225
                                    },
                                    "track_number": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "lanzamiento_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Canción creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "lanzamientoId": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "lanzamiento": {
                                                    "type": "string",
                                                    "example": "Nombre del lanzamiento",
                                                    "nullable": true
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "Mi canción"
                                                },
                                                "duracion": {
                                                    "type": "integer",
                                                    "example": 225
                                                },
                                                "duracionFormateada": {
                                                    "type": "string",
                                                    "example": "03:45"
                                                },
                                                "trackNumber": {
                                                    "type": "integer",
                                                    "example": 1,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "titulo": [
                                                    "El título de la canción es obligatorio."
                                                ],
                                                "lanzamiento_id": [
                                                    "El lanzamiento asociado no existe."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Introduce: Bearer {token}"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}